Convert SessionContext to a class
