+++ testing/trino-tests/src/test/java/io/trino/tests/TestMetadataManager.java
@@ -140,7 +140,7 @@ public class TestMetadataManager
         dispatchManager.createQuery(
                 queryId,
                 Slug.createNew(),
-                new TestingSessionContext(TEST_SESSION),
+                TestingSessionContext.fromSession(TEST_SESSION),
                 "SELECT * FROM lineitem")
                 .get();
